+++ js/data/timer.test.js
@@ -1,20 +1,5 @@
 import {assert} from 'chai';
-
-// Функция-конструктор для генерации объекта таймера
-function CountSeconds(time) {
-  let count = time;
-  this.tick = function () {
-    const value = count - 1;
-
-    if (count <= 0) {
-      throw new Error(`The timer expired`);
-    } else {
-      count -= 1;
-    }
-
-    return value;
-  };
-}
+import CountSeconds from './../timer.js';
 
 let timer = new CountSeconds(30);
